// web/@types/appConfig.d.ts
export interface SystemDataType {
    aliasNum: string;
    country: string;
    appUrl: string;
    androidAppUrl: string;
    cdn: string;
    siteName: string;
    seoDescription: string;
    seoKey: string;
    siteLogo: string;
    pcLogo: string;
    doMainUrl: string;
    version: string;
    advertisingSpace: string;
    ico: string;
    appNoInvitationCode: string;
    kfUrl: string;
    iosQr: string;
    androidQr: string;
    wapDomain: string;
    appDomain: string;
    timestamp: number;
    isUseAppUrl: boolean;
    isOpenNewCustomer: boolean;
    isUseAndroidAppUrl: boolean;
    isOpenPlatformLottery: string;
    popTime: number;
    indexShowTab: string;
    CaptchaAppId: null;
    repairToolsUrl: string;
    isRechargeBindBankCard: boolean;
    cgpUsdtRmbRate: any;
    usdtPayTutorialsAddress: string;
    isPollingPayBank: boolean;
    currency: string;
    defaultAreaCode: string;
    toDayWithdrawalUpper: string;
    toDayWithdrawalTimesUpper: string;
    withdrawalOwer: string;
    withdrawalUpper: string;
    sportBackground: string;
    customer_background: string;
    seoTitle: string;
    isOpenPromoPreferential: boolean;
    alias_num: string;
}